pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services/negotiation"]

[project]
name = "samson-negotiation-service"
version = "0.1.0"
description = "FastAPI service for diplomatic negotiations with WebRTC A/V streaming"
authors = [
    {name = "Samson Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "aiortc>=1.5.0",
    "pydantic>=2.5.0",
    "structlog>=23.2.0",
    "websockets>=12.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "ruamel.yaml>=0.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "black>=23.10.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.13.1",
    "detect-secrets>=1.4.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
